feat: Add progress tracking for data fetching
Introduced a new feature that allow the users to see the progress of
downloading, uploading,fetching data, etc.

- Adding a progress tracking mechanism using `onDownloadProgress` in
axios requests.
- Updating `FactoryLogContextProvider` to include progress logging
during data fetch.
- Modifying `FactoryLogPreFilter` to show a progress bar and loading
state during data fetching.(Not working)
- Creating a new API utility file `factoryApi.ts` to handle data
fetching with progress tracking.

These changes are currently in testing. There are a lot of repeating
codes. The whole process is not very streamline. I might remove the
`FactoryLogContext.tsx` in the future or shrink it down.