refactor: Enhance useCloseAndEscape hook
- Added a new parameter `returnFocusOnClose` to the `useCloseAndEscape`
- `useCloseAndEscape` control whether the focus returns to the trigger
element after closing.
- Updated ToggleHiddenRow component to use `returnFocusOnClose` set
to `false` (focus does not return to trigger).
- This change resolves focus management issues where different
components required different focus behavior when closed.

`useCloseAndEscape` automatically make the last focused element regain
focus when a modal is closed. When this hook is used within
`ToggleHiddenRow` component, the focus state does not return to the
correct button after closing. Instead, it consistently returns to the
first button that was interacted with (because that button is the last
focused element). Regardless of which button trigger the most recent
action. User need to click outside of the button twice for it to lose
the focus state (first is to close the dropdown, the second is to
make the button lose focus.) This changes allow this behavior to be turn
off if needed.