feat: Add disabled state support to SelectField
When the state of the mold is not included in the predefined options,
the `select` component is disabled because it's not intended to be
modified by the user. Key changes included:

- Passing a `disabled` prop to the `SelectField` component in
`Update.tsx`.
- Modifying the `SelectField` to conditionally disable the field based
on the `disabled` prop.
- Updating the `useUpdate` hook to check if the `state` exists in the
predefined options (`isStateInPredefinedOptions`) and use that to
disable the `state` dropdown.

This ensures better control over the `state` selection and prevents
unintended modifications when certain options are not available.