feat: Add form validation
- Added required fields and input validation patterns for `CreateMold`
 and `CreateMoldSubComponent` forms.
- Enhanced UX by providing visual feedback on input validation errors,
such as highlighting invalid fields in red.
- Improved form handling and submission logic by introducing validation
checks and resetting invalid state after user input changes.

Visual feedback for validation error in Select component is not behaving
as intended. The red border is applied to all Select with invalid value
upon submitting, not just the first one. The required message is only
shown when form is submitted for the first time.