refactor: Reintroduce useCreateMold to CreateMold
- Added a new function `getNewSnNum` to fetch `sn_num`. This function
can handle both normal cases and special case(when sys is 雙色系列). It
should work for 模仁 with some tweaking.
- Introduced new state `userIsStillEditing` in `useCreateMold` to check
if user is still entering data. This is used as one of the conditions
for when useQuery should fetch new `sn_num`.
- Introduced three new states that are related to fetching `sn_num`.
(`mold_num`, `sys`, and `sn_target`)
- Added a new function `clearForm` in `useCreateMold` to not only clear
- Reintroduced `useCreateMold` hook in `CreateMold.tsx` to manage state
and functionality related fetching `sn_num`
- Refacotored `CreateMold.tsx` to utilize `useCreateMold` hook.