fix: cache update logic for mold data
1. Changed queryKey in `usePreFilter` from `["preFilterData"]` to
`["preFilterData", params]`. Otherwise, the data will not be updated.

2.  Replaced `data` with `currentMoldData` in the `Update` and `useUpdate`
components for clarity and consistency.

3. Adjusted the return value from `usePreFilter` from `response.data.data`
(`FilterData["data"]`) to `response.data` to expose post property,
which is necessary to differentiate whether to update, delete, add, or
do nothing when updating individual cache.

4. Adjusted `useCreateMold` and `useUpdate` hooks to use `setQueriesData`
for better query cache handling.

5. Fixed the logic for updating, removing, or adding mold data to the
query cache based on whether the mold was present before and after the
update.

6. Modified the `usePreFilter` hook to correctly handle `data` within
the query response and ensure the appropriate data structure is used.

After changing query in `usePreFilter`, instead of one cache, now we
have multiple cache for different params. So the update logic in
`useUpdate` also need a update to handle different cases. To identify
whether the updated mold data is in the particulart cache, I need access
to params, hence the third changes.

This fix ensures the query cache is updated properly based on mold data
changes and uses clearer naming conventions for better code readability.