fix: improve type handling in FormSelectField
**Description**:
- Updated `FormSelectFieldProps` to enhance the handling of `value` and
`onChange` props with conditional typing. (`value` and `onChange` props
are always required in previous type definition which is not how it was
intended)
- The `value` and `onChange` props now follow a more robust and clear
distinction between controlled and uncontrolled states.
- Refined the `onChange` handler to correctly pass the selected `value`
(`e.value`) when updating the form field and calling external `onChange`
logic if provided.
- Enhanced the value prioritization logic to default to the field's
state unless overridden by a provided `value`.